pyproject.toml
[tool.poetry]
name = "diplomaworkproject"
version = "0.1.0"
description = ""
authors = ["Drapeyko Nikita <drapeykonik@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.11"
pdoc3 = "^0.10.0"
joblib = "^1.2.0"
interrogate = "^1.5.0"
notebook = "^6.5.4"
progressbar = "^2.5"
scikit-learn = "^1.2.2"
pandas = "^2.0.1"
pydantic = "^1.10.7"
scikit-image = "^0.20.0"
pillow = "^9.5.0"
tqdm = "^4.65.0"
matplotlib = "^3.7.1"
tensorboard = "^2.13.0"
openmim = "^0.3.7"
streamlit = "^1.22.0"
torch = { url = "https://download.pytorch.org/whl/cu117/torch-2.0.1%2Bcu117-cp310-cp310-linux_x86_64.whl"}
torchvision = { url = "https://download.pytorch.org/whl/cu117/torchvision-0.15.2%2Bcu117-cp310-cp310-linux_x86_64.whl"}
uvicorn = "^0.22.0"
fastapi = "^0.95.2"
python-multipart = "^0.0.6"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
isort = "^5.12.0"
pre-commit = "^3.2.2"
pytest = "^7.3.1"
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
	\.git
| \.hg
| \.mypy_cache
| \.tox
| \.venv
| _build
| buck-out
| build   
)/ 
'''

[tool.interrogate]
exclude = ["src/ssd/mmdetection/"]
